Added functionality for timestamp loading and saving

+++ core/Timer.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-import os, shutil
+import os, shutil, json
 
 
 class Timer(object):
@@ -46,18 +46,28 @@ class Timer(object):
         return os.path.isdir(cls.getFolderPath())
     
 
-    def createFolder(cls):
+    def createFolder(cls) -> bool:
         """Attempts to create a folder for timestamps
+
+        Returns:
+            bool: Success?
         """
         if not cls.folderExists():
             os.mkdir(cls.getFolderPath())
+            return True
+        return False
     
 
-    def deleteFolder(cls):
+    def deleteFolder(cls) -> bool:
         """Attempts to delete timestamp folder with files inside
+
+        Returns:
+            bool: Success?
         """
         if cls.folderExists():
             shutil.rmtree(cls.getFolderPath())
+            return True
+        return False
     
 
     #########################################################################################
@@ -81,19 +91,66 @@ class Timer(object):
         return os.path.isfile(cls.getFilePath())
     
 
-    def createFile(cls):
+    def createFile(cls) -> bool:
         """Create a file for timestamps
+
+        Returns:
+            bool: Success?
         """
+        if not cls.folderExists():
+            cls.createFolder()
         if not cls.fileExists():
             with open(cls.getFilePath(), 'w'):
                 pass
+            return True
+        return False
     
 
-    def deleteFile(cls):
+    def deleteFile(cls) -> bool:
         """Removing timestamp file
+
+        Returns:
+            bool: Success?
         """
+        if not cls.folderExists():
+            return False
         if cls.fileExists():
             os.remove(cls.getFilePath())
+            return True
+        else:
+            return False
+    
+
+    #########################################################################################
+
+
+    def loadTimestamps(cls) -> list:
+        """Attempts to load list of timestamps
+
+        Returns:
+            list: List of timestamps
+        """
+        if not cls.folderExists():
+            cls.createFolder()
+        if cls.fileExists():
+            return json.load(open(cls.getFilePath(), "r"))
+        else:
+            return []
+
+
+    def saveTimestamps(cls, timestamps: list) -> bool:
+        """Attempts to save list of timestamps
+
+        Args:
+            timestamps (list): List of timestamps to save
+
+        Returns:
+            bool: Success?
+        """
+        if not cls.folderExists():
+            cls.createFolder()
+        json.dump(timestamps, open(cls.getFilePath(), "w"), indent=4, sort_keys=False)
+        return True
     
 
     #########################################################################################
